Remove old index/socket; rename protos
