+++ package.json
@@ -3,8 +3,8 @@
   "version": "1.0.0",
   "main": "src/index.mjs",
   "scripts": {
-    "start": "node --experimental-modules src/protoindex.mjs",
-    "dbg": "node --experimental-modules --inspect src/protoindex.mjs"
+    "start": "node --experimental-modules src/index.mjs",
+    "dbg": "node --experimental-modules --inspect src/index.mjs"
   },
   "engines": {
     "node": ">=12"
